Starting to refactor main module.
