+++ .travis.yml
@@ -3,8 +3,6 @@ sudo: false
 python:
   - '2.6'
   - '2.7'
-  - '3.3'
-  - '3.4'
   - pypy
 deploy:
   provider: pypi
