+++ TODO.md
@@ -1,0 +1,15 @@
+Todo list for Google-Search-API
+====
+
+## Next tasks
+
+1. Mock html test for search images
+2. Refactor and split main modules in separated ones
+3. Write tests for future methods development
+
+## Stuff to check out
+
+* vcrpy to record web requests and make mocks
+* SauceLabs to do UI tests
+* Sphinx para generar la documentación - estilo readthedocs
+* Cookiecutter-pythonpackage
